Revert "Replace Ibanomat with Ibanizator"
The BIC lookup will not return the correct string.

This is due to how the BIC lookup works in the Ibanizator gem.
It does its lookup via a bank code (blz) but the problem is that
more than one bank can have the same blz. When the gem populates
the list of blzs it will only store the last blz with the corresponding
bank information (BIC, bank name, address and so on..).

In order to do a proper lookup, you will probably need the address of the
bank as well.

This reverts commit 3f6cddfc9b9164bc005dbe2293ce69052fb92dd9.