Previous release of the gem broke the API
The reason for that is that we now calculate/lookup the IBAN/BIC
locally. Meaning some of the exceptions has been changed.

BankTools::DE::IbanBicConverter::ServiceUnavailable and
BankTools::DE::IbanBicConverter::UnknownError is removed since we're not
using a service anymore.

BankTools::DE::IbanBicConverter::CouldNotConvertError is replaced with
BankTools::DE::IbanBicConverter::CouldNotConvertIbanError and
BankTools::DE::IbanBicConverter::CouldNotFindBicError.